Start work on API access and routes.
Implemented by having a different factory for the api end points, and by
mounting the API end point under /api in the application.

The API view class retrieves the base site from the api key, requiring a
valid key to allow the request to proceed. We therefore assign the 'api'
access right to everyone for now, but this could probably change to a more
dynamic access right later with a authenticated user id mapped to a site
key or something similar.

This should however work, so that's good enough for now.