Implement re-orderable recursive layouts. Ick.
Well, it works. It isn't pretty. It will need a rewrite in the future.
This is bad.

But it works. So we'll go with it for now, until we decide to rewrite the
whole layout engine for the page editing. Which we'll do when we know what
we actually need.

For now, we continue our prototype.